requirements.txt
Flask==3.1.2
gunicorn==23.0.0
dash==3.2.0
dash-bootstrap-components==2.0.4
pandas==2.3.2
plotly==6.3.0
numpy==2.2.6
scikit-learn==1.7.2
scipy==1.15.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
spotipy==2.25.1
cryptography==46.0.1
redis==6.4.0
Werkzeug==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
urllib3==2.5.0
typing_extensions==4.15.0
psycopg2-binary==2.9.11
bcrypt
# pipreqs==0.5.0

